Trait pallet_asset_registry::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RegistryOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type AssetId: Parameter + Member + Default + Copy + BaseArithmetic + MaybeSerializeDeserialize + MaxEncodedLen;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen;
    type AssetNativeLocation: Parameter + Member + Default + MaxEncodedLen;
    type StringLimit: Get<u32>;
    type MaxAssetIdInPool: Get<u32> + Clone + Debug + TypeInfo + Eq + PartialEq;
    type NativeAssetId: Get<Self::AssetId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type RegistryOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which can work with asset-registry.

type AssetId: Parameter + Member + Default + Copy + BaseArithmetic + MaybeSerializeDeserialize + MaxEncodedLen

Asset type

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen

Balance type

type AssetNativeLocation: Parameter + Member + Default + MaxEncodedLen

Asset location type

type StringLimit: Get<u32>

The maximum length of a name or symbol stored on-chain.

type MaxAssetIdInPool: Get<u32> + Clone + Debug + TypeInfo + Eq + PartialEq

The maximum number of assets in a pool

type NativeAssetId: Get<Self::AssetId>

Native Asset Id

type WeightInfo: WeightInfo

Weight information for the extrinsics