Trait pallet_asset_registry::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type RegistryOrigin: EnsureOrigin<Self::Origin>;
    type AssetId: Parameter + Member + Default + Copy + BaseArithmetic + MaybeSerializeDeserialize + MaxEncodedLen;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize;
    type AssetNativeLocation: Parameter + Member + Default;
    type StringLimit: Get<u32>;
    type NativeAssetId: Get<Self::AssetId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type RegistryOrigin: EnsureOrigin<Self::Origin>

The origin which can work with asset-registry.

type AssetId: Parameter + Member + Default + Copy + BaseArithmetic + MaybeSerializeDeserialize + MaxEncodedLen

Asset type

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize

Balance type

type AssetNativeLocation: Parameter + Member + Default

Asset location type

type StringLimit: Get<u32>

The maximum length of a name or symbol stored on-chain.

type NativeAssetId: Get<Self::AssetId>

Native Asset Id

type WeightInfo: WeightInfo

Weight information for the extrinsics